requirements.txt
django_jalali_date==0.2.10
persiantools==1.4.1
jdatetime==3.2.0
social_auth_app_django==3.1.0
pytz==2019.1
Django==2.2.2
django_crispy_forms==1.7.2
python_dateutil==2.8.0
